Add files via upload
This was a university project in which a neural network was created and trained to predict Amazonian deforestation. The network was first trained with real data, then tested with a cluster of data that was not used in the training  to ensure reliability in the prediction and finally was used to predict future deforestation. 
The neural network reliability was also compared with other forecasting methods, for example, the polynomial regression.